fix: separate two different task
